+++ client/components/TableOfContentsMarkdownDesktop.vue
@@ -1,15 +1,16 @@
 <template>
-  <div
-    class="hidden lg:sticky lg:top-[3px] lg:flex lg:flex-col lg:min-w-[300px] p-2"
-  >
+  <div class="hidden lg:sticky lg:top-[0px] lg:h-screen lg:flex lg:flex-col lg:min-w-[300px]">
     <TableOfContentsHighlight
       :toc="props.toc"
       list-type="ordered"
-      wrapper-class="h-[calc(100vh-24px)] pb-2"
+      wrapper-class="flex flex-col min-h-0 pt-4 pb-2 px-4"
       list-class="mr-1"
-      nested-list-class="pl-1"
-      :list-item-class="`block text-sm py-2 border-t-1 border-t-gray-300 dark:border-t-gray-500 no-underline hover:underline ${ANCHOR_TAILWIND_STYLE}`"
-      list-item-active-class="text-shadow-bold"
+      nested-list-class="pl-2"
+      :links-class="`block text-sm py-2 dark:border-t-gray-500 ${ANCHOR_TAILWIND_STYLE}`"
+      links-active-class="text-shadow-bold"
+      link-class="block no-underline hover:underline"
+      last-link-class="flex-1 flex"
+      :show-last-link-icon="true"
     >
       <Heading
         level="2"
