+++ client/pages/about/[...slug].vue
@@ -1,0 +1,3 @@
+<template>
+  <MarkdownPage />
+</template>
